Move tree making to separate module
